V1.2: Rewrite major parts for multi-agent support
This version modifies major parts of the environment to enable
compatibility with MARL scenarios implemented in pettingzoo. Two new
scenarios are implemented, a tracking scenario based on the existing
racing scenario where the environment is perceived as a trajectory, and
a MARL scenario where agents must cooperatively navigate an unregulated
four-way intersection.